Updated microsoft_face
